fix:修复就近路由标签无法下发问题
